always set bls_pubkey and fund vote accounts in tests/helpers
