+++ ledger/src/staking_utils.rs
@@ -1,8 +1,10 @@
 #[cfg(test)]
 pub(crate) mod tests {
     use {
+        agave_votor_messages::consensus_message::BLS_KEYPAIR_DERIVE_SEED,
         rand::Rng,
         solana_account::AccountSharedData,
+        solana_bls_signatures::keypair::Keypair as BLSKeypair,
         solana_clock::Clock,
         solana_instruction::Instruction,
         solana_keypair::Keypair,
@@ -18,7 +20,10 @@ pub(crate) mod tests {
         solana_vote::vote_account::{VoteAccount, VoteAccounts},
         solana_vote_program::{
             vote_instruction,
-            vote_state::{VoteInit, VoteStateV4, VoteStateVersions},
+            vote_state::{
+                VoteAuthorize, VoteInit, VoteStateV4, VoteStateVersions, VoterWithBLSArgs,
+                create_bls_proof_of_possession,
+            },
         },
         std::sync::Arc,
     };
@@ -59,6 +64,27 @@ pub(crate) mod tests {
                     ..vote_instruction::CreateVoteAccountConfig::default()
                 },
             ),
+        );
+
+        // Add BLS pubkey to the vote account using the authorize instruction with BLS.
+        // This sets the authorized voter to the same pubkey but adds the BLS key.
+        let bls_keypair =
+            BLSKeypair::derive_from_signer(vote_account, BLS_KEYPAIR_DERIVE_SEED).unwrap();
+        let (bls_pubkey, bls_proof_of_possession) =
+            create_bls_proof_of_possession(&vote_pubkey, &bls_keypair);
+
+        process_instructions(
+            bank,
+            &[from_account, vote_account],
+            &[vote_instruction::authorize(
+                &vote_pubkey,
+                &vote_pubkey, // currently authorized voter
+                &vote_pubkey, // new authorized voter (same, just adding BLS)
+                VoteAuthorize::VoterWithBLS(VoterWithBLSArgs {
+                    bls_pubkey,
+                    bls_proof_of_possession,
+                }),
+            )],
         );
 
         let stake_account_keypair = Keypair::new();
